Refactor and format
